+++ packages/expo/CHANGELOG.md
@@ -19,6 +19,7 @@
 
 - [android] Add getDefaultReactHost to ExpoReactHostFactory ([#40086](https://github.com/expo/expo/pull/40086) by [@gabrieldonadel](https://github.com/gabrieldonadel))
 - [Android] Remove edge-to-edge logic from `ReactActivityDelegateWrapper`. ([#40263](https://github.com/expo/expo/pull/40263) by [@behenate](https://github.com/behenate))
+- [internal] Add `source` to `DOMPropsInternal` to enable pre-bundled DOM Components.
 
 ### ⚠️ Notices